fix: split oversized backend files for backend-size CI
Move factory prompt run CLI tests and handlingBehavior validator rules into
dedicated files so root_run_test.go, config_validator.go, and
config_validator_test.go stay within the 1000-line backend-size limit.
